// src/spotify.types.ts
export type CurrentlyPlayingTrack = {
	device: {
		id: string;
		is_active: boolean;
		is_private_session: boolean;
		is_restricted: boolean;
		name: string;
		type: string;
		volume_percent: number;
		supports_volume: boolean;
	};
	repeat_state: string;
	shuffle_state: boolean;
	context: {
		type: string;
		href: string;
		external_urls: {
			spotify: string;
		};
		uri: string;
	};
	timestamp: number;
	progress_ms: number;
	is_playing: boolean;
	item: {
		album: {
			album_type: string;
			total_tracks: number;
			available_markets: string[];
			external_urls: {
				spotify: string;
			};
			href: string;
			id: string;
			images: [
				{
					url: string;
					height: number;
					width: number;
				}
			];
			name: string;
			release_date: string;
			release_date_precision: string;
			restrictions: {
				reason: string;
			};
			type: string;
			uri: string;
			artists: [
				{
					external_urls: {
						spotify: string;
					};
					href: string;
					id: string;
					name: string;
					type: string;
					uri: string;
				}
			];
		};
		artists: [
			{
				external_urls: {
					spotify: string;
				};
				followers: {
					href: string;
					total: number;
				};
				genres: string[];
				href: string;
				id: string;
				images: [
					{
						url: string;
						height: number;
						width: number;
					}
				];
				name: string;
				popularity: 0;
				type: string;
				uri: string;
			}
		];
		available_markets: [string];
		disc_number: string;
		duration_ms: string;
		explicit: boolean;
		external_ids: {
			isrc: string;
			ean: string;
			upc: string;
		};
		external_urls: {
			spotify: string;
		};
		href: string;
		id: string;
		is_playable: boolean;
		linked_from: {};
		restrictions: {
			reason: string;
		};
		name: string;
		popularity: number;
		preview_url: string;
		track_number: number;
		type: string;
		uri: string;
		is_local: boolean;
	};
	currently_playing_type: string;
	actions: {
		interrupting_playback: boolean;
		pausing: boolean;
		resuming: boolean;
		seeking: boolean;
		skipping_next: boolean;
		skipping_prev: boolean;
		toggling_repeat_context: boolean;
		toggling_shuffle: boolean;
		toggling_repeat_track: boolean;
		transferring_playback: boolean;
	};
};

export type AccessTokenResponse = {
	access_token: string;
	token_type: string;
	scope: string;
	expires_in: number;
	refresh_token: string;
};

export type SpotifyAuthCallback = {
	action: string;
	error?: string;
	code?: string;
	state: string;
};